rag_web_backend/pyproject.toml
[project]
name = "rag-web-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0,<25.0.0",
    "alembic>=1.13.0,<2.0.0",
    "asyncpg>=0.29.0,<0.30.0",
    "bcrypt>=4.0.1,<5.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<0.116.0",
    "markitdown[all]>=0.0.1a2",
    "opencv-python-headless>=4.11.0.86",
    "mineru[core]>=2.1.11",
    "numpy>=1.24.0",
    "opencc-python-reimplemented>=0.1.7",
    "openpyxl>=3.1.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "pydantic>=2.9.0,<3.0.0",
    "pydantic-settings>=2.5.0,<3.0.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "python-multipart>=0.0.9,<0.1.0",
    "redis>=5.0.0,<6.0.0",
    "requests>=2.31.0",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.35,<3.0.0",
    "tqdm>=4.66.0",
    "uvicorn[standard]>=0.30.0,<0.31.0",
]
